Files 4.0 (#87)
Files 4.0

This change includes a brand new implementation for Files,
that keeps most of the public API intact, while modernizing
the underlying implementation.

Under the hood, Files now uses value types and protocols,
rather than class inheritance - which improves the type safety
of the API, and streamlines the underlying code. Error handling
is also improved to include more underlying info, and the docs
have been reworked to be much more thorough.

The API has also been fine-tuned and modernized, dropping the
`FileSystem` class in favor of more “Swifty” APIs on `Folder`,
introducing more options for creating new files and folders, and
making it possible to easily change properties on file system
sequences.

A list of all changes will be posted as part of this version’s
release notes.