Folder: Add API for resolving search paths (#112)
This change adds a public API for resolving a folder for a given search
path within a domain. This new API is now also used internally to resolve
the current user’s “Documents” and “Library” folders. It's only available
on Apple's platforms (for now), since the underlying method that it uses
is currently not fully implemented on other platforms.